During uri normalisation, make use of pre-parsed port_number
If the uri has already had it's port parsed and converted to int,
for example if it is a CONNECT and is parsed with:
  htp_parse_uri_hostport
then the validated and converted port number is already available
to set in the normalised uri structure.
At least, this is a nice optimisation, at best, it fixes the
problem of the normalised uri structure not containing any
port_number value at all, when in fact, the raw uri did have one.