Chunked Header Length
Add a field: tx_detail::{request,response}_chunk_header_len, that
records the length of chunked headers in bytes.  The host
application can use this figure to seek the data stream, past the
chunked headers to the beginning of the data proper, thus
isolating the real data, for example for scanning or logging
purposes.

Pieces of the stream that are considered part of the chunk header
are:
- The block of hexidecimal characters that form the chunk length
- The line terminator at the end of the chunk length
- The line terminator at the end of the chunk data

The hook where the chunk_header_len is relevant is the first
body data hook in both the request and response.  After this hook
occasion, the chunk_header_len will be zerod so as to avoid
double accounting, and to prepare for the next data chunk header.