bstr_safe_assign - assign bstr* with auto freeing and error checking
When using this function to assign bstr* to bstr*, the target bstr*
will automatically be checked for non null value and freed if so.
If used in a debug mode, will also assert to help find occurences
of such double assignments with the aim of eventually eliminating
this situation as it is un-natural in most cases.